QStyleSheet/QToolbox: don't overlap text and icon
When an icon was specified with a style sheet, the content rect for the
toolbox label was not adjusted which results in an overlapping of the
text and the icon.
Fix it by adjusting the rect used by drawing the label when an icon is
specified with a style sheet.

Fixes: QTBUG-40871
Change-Id: I088b319e6355d78066d0bc4472882a5d9a599091
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>
Reviewed-by: Vitaly Fanaskov <vitaly.fanaskov@qt.io>