Update and expand public suffix list's attribution information
Say where the git repo is, for finding commit IDs, but make clear that
the download should come from the public site.  Revised Description to
the closest match to it I could find on the current web-site (on the
home page).  Mention where the script to convert the data live.

Task-number: QTBUG-72623
Change-Id: Ie5f7b2b0c21cdf1c24e311c13866cb1bb02e6973
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>